moved the OracleHealthTransitionAlert.jsx to the shared folder
