+++ src/applications/mhv-medications/components/MedicationsList/MedicationsListCard.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom-v5-compat';
 import { useSelector } from 'react-redux';
 import ExtraDetails from '../shared/ExtraDetails';
 import LastFilledInfo from '../shared/LastFilledInfo';
-import { OracleHealthInCardAlert } from '../OracleHealthTransitionAlerts';
+import { OracleHealthInCardAlert } from '../shared/OracleHealthTransitionAlerts';
 import {
   dateFormat,
   getPrescriptionDetailUrl,
