+++ src/applications/mhv-medications/containers/RefillPrescriptions.jsx
@@ -52,7 +52,7 @@ import {
   selectCernerPilotFlag,
   selectMhvMedicationsOracleHealthCutoverFlag,
 } from '../util/selectors';
-import { OracleHealthT3Alert } from '../components/OracleHealthTransitionAlerts';
+import { OracleHealthT3Alert } from '../components/shared/OracleHealthTransitionAlerts';
 import { filterPrescriptionsByTransition } from '../util/oracleHealthTransition';
 
 import { selectSortOption } from '../selectors/selectPreferences';
